Move the PID stuff into an auton action
